git-svn-id: http://ctp2.darkdust.net/anonsvn/branches/port-trunk@1073 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/ctpai.cpp
@@ -1538,8 +1538,8 @@ void CtpAi::MakeRoomForNewUnits(const PLAYER_INDEX playerId)
 
 						sint16 defense_count;
 						sint16 tmp_count;
-						double tmp;
-						double defense_strength;
+						float  tmp;
+						float defense_strength;
 						move_army->ComputeStrength(tmp,
 						                           defense_strength,
 						                           tmp,
@@ -1942,8 +1942,8 @@ void CtpAi::ComputeCityGarrisons(const PLAYER_INDEX playerId )
 
 		sint16 defense_count;
 		sint16 tmp_count;
-		double tmp;
-		double defense_strength;
+		float  tmp;
+		float  defense_strength;
 		army->ComputeStrength(tmp,
 		                      defense_strength,
 		                      tmp,
